fix: Fixed attribute 2 (fortify)
Prevented attribute 2 from giving defense bonuses to troops which should not have had defense bonuses. This bug used to occur when a game was loaded, since fortify would automatically add the bonus whenever the attribute code was loaded. Now, the attribute checks if the unit is really on a city before adding a bonus